Create DeferredCommand.java
A deferred command allows individuals to run the command only when scheduled instead of prior to it. A use case is when teams attempt to build a new trajectory using data from vision in the middle of autonomous. Building a trajectory otherwise (in an InstantCommand for example) would build with a vector (0,0). Deferring the command allows it to build correctly.